==> code/10_visualize_new_data_conservative.Rmd <==
Analyze data to get a number on prescribed burns in the Western US & visualize comparisons to low-severity burns
This script is the most recent
Tyler L. McIntosh
CU Boulder CIRES Earth Lab
Last updated: 2/8/23

This script uses the following naming conventions wherever possible:
 lowerCamelCase for variables
 period.separated for functions
 underscore_separated for files

# Setup

```{r, message = FALSE, warning = FALSE, include = FALSE}
## Libraries ----

#Check the required libraries and download if needed
list.of.packages <- c("tidyverse", #Includes ggplot2, dplyr, tidyr, readr, purrr, tibble, stringr, forcats
                      "terra",
                      "sf",
                      "mapview",
                      "here",
                      "future", "future.apply", "furrr", "doFuture", "progressr", #Futureverse!
                      "tictoc", 
                      "mblm", #Median-based linear models (i.e. thiel-sen)
                      "plyr",
                      "gridExtra", #complex graphics
                      "kableExtra", #nice tables
                      "magick", #for kableExtra
                      "googlesheets4", #read in google sheet data
                      "scales", #add commas to ggplot axis
                      "tigris", #US data
                      "tmap", #mapping
                      "webshot2",
                      "webshot",
                      "patchwork",
                      "scales") #add commas to axis

#Install all packages that aren't installed yet
install.packages(setdiff(list.of.packages, rownames(installed.packages())))

#Load all packages
invisible(lapply(list.of.packages, library, character.only = TRUE)) #use 'invisible' to prevent output being displayed

## Clean workspace & set up environment ----
rm(list=ls()) #Ensure empty workspace if running from beginning
here::i_am("code/10_visualize_new_data_conservative.Rmd")
here::here() #Check here location
#OR
#setwd() #Set working directory directly
options(digits = 6) #Set standard decimal print output
options(scipen = 999) #Turn scientific notation on and off (0 = on, 999 = off)

#Ensure necessary directories are created
figsDir <- here::here("figs")
if (!dir.exists(figsDir)){
  dir.create(figsDir)
}
figsDir <- here::here("figs/both_lcms_lcmap")
if (!dir.exists(figsDir)){
  dir.create(figsDir)
}


derivedDatDir <- here::here("data", "derived")
if (!dir.exists(derivedDatDir)){
  dir.create(derivedDatDir)
}

```

# Read in data
Read in the data;
-From Rud
-From rx burn data munging & collating


```{r}
#From script 02
nfporsWithGee <- read_csv(here::here(derivedDatDir, "nfpors_with_gee_lcms_lcmap.csv"))

#NEW data summaries from Rud, includes small fires
yearSummary <- googlesheets4::read_sheet("https://docs.google.com/spreadsheets/d/1wUWLin5cXSd9uflna9f_hliz5HEWxOcc7a58mtDziEk/edit#gid=1307905708", sheet = "Year Summary", range = "A1:J12")
stateSummary <- googlesheets4::read_sheet("https://docs.google.com/spreadsheets/d/1wUWLin5cXSd9uflna9f_hliz5HEWxOcc7a58mtDziEk/edit#gid=1307905708", sheet = "State Summary", range = "A1:K12")

#OLD data from Rud, events
mtbs <- googlesheets4::read_sheet("https://docs.google.com/spreadsheets/d/1cGDeCfF6Gzkf9P4SRHDfqPmZfwyX438jURyzrL49IGk/edit?pli=1#gid=0", sheet = "MTBS Fire Events")
smallFires <- googlesheets4::read_sheet("https://docs.google.com/spreadsheets/d/1cGDeCfF6Gzkf9P4SRHDfqPmZfwyX438jURyzrL49IGk/edit?pli=1#gid=0", sheet = "Small Fire Events")


#Remove new unburned column
yearSummary <- yearSummary |> dplyr::select(-Unburned)
stateSummary <- stateSummary |> dplyr::select(-Unburned)


#Pull in conservative forest stats, from GEE script conservative_forest_stats.js
fPixFiles <- list.files(here::here('data/derived'), pattern = "forest_pixel_count", full.names = TRUE)
fPixAll <- fPixFiles %>% 
  purrr::map_dfr(~read_csv(.)) |>
  dplyr::mutate(hectares = conservativeForest * 900 * 0.0001) |> #conversion from 30x30m (900m^2) pixels to hectares
  dplyr::select(NAME, year, hectares)


```


# Summary stats
```{r}

# Function to fix a current (3/20/24) issue with kableExtra::save_kable in which it will not export .pngs
# PARAMETERS
# k : output from kable
# filePath : the full desired filepath (e.g. 'myDir/figs/myTable.png')
save_kable_workaround <- function(k, filePath) {
  htmlPath <- paste0(tools::file_path_sans_ext(filePath), ".html")
  kableExtra::save_kable(x = k, file = htmlPath)
  webshot2::webshot(htmlPath, file = filePath)
  file.remove(htmlPath)
}


#Filter to just forest rx up to 2020
nfporsWithGee |>
  dplyr::filter(ACTUALCOMPLETIONYEARNEW < 2021)


#Summary stat dataset versions

nfporsWithGeeForest <- nfporsWithGee |>
  dplyr::filter(LCMS_PreBurnYear == 1 & LCMAP_PreBurnYear == 4)

nfporsWithGeeForestBroadcast <- nfporsWithGeeForest |> dplyr::filter(TYPENAME == "Broadcast Burn")

nfporsWithGeeForestFRGLowMix = dplyr::filter(nfporsWithGeeForest, FRGDescription == "frcLowMix")

nfporsWithGeeForestBroadcastFRG <- nfporsWithGeeForestFRGLowMix |> dplyr::filter(TYPENAME == "Broadcast Burn")


mtbsForest <- mtbs |>
  dplyr::mutate(percFor = preFireFor / areaHA * 100) |>
  dplyr::filter(percFor >= 10)

smallFiresForest <- smallFires |>
  dplyr::mutate(percFor = preFireFor / GIS_Hectar * 100) |>
  dplyr::filter(percFor >= 10)

mtbsForestLowFRG <- mtbs |>
  dplyr::mutate(forFrgLow = frcLowCbiHigh + frcLowCbiLow + frcLowCbiUnb,
                percForLowFRG = forFrgLow / areaHA * 100) |>
  dplyr::filter(percForLowFRG >= 10)

# smallFiresForestLowFRG <- smallFires |>
#   dplyr::mutate(forFrgLow = frcLowCbiHigh + frcLowCbiLow + frcLowCbiUnb,
#                 percForLowFRG = forFrgLow / areaHA * 100) |>
#   dplyr::filter(percForLowFRG >= 10)

#Summary stats
totHighGood <- sum(yearSummary$`High Goodfire (CBI High FRG Replacement, HA)`)
totLowGood <- sum(yearSummary$`Low/Mod Goodfire (CBI Low/mod in FRG low/mod, HA)`)
totFire <- sum(mtbs$areaHA) + sum(smallFires$GIS_Hectar)
totForFire <- sum(yearSummary$`burnFor (low+mod+high CBI, HA)`)
totRx <- sum(nfporsWithGee$TOTALACCOMPLISHMENT_HA)
totRxForest <- sum(nfporsWithGeeForest$TOTALACCOMPLISHMENT_HA)
totRxForestFRG <- sum(nfporsWithGeeForestFRGLowMix$TOTALACCOMPLISHMENT_HA)
totRxForestBroadcast <- sum(nfporsWithGeeForestBroadcast$TOTALACCOMPLISHMENT_HA)
totRxForestBroadcastFRG <- sum(nfporsWithGeeForestBroadcastFRG$TOTALACCOMPLISHMENT_HA)
totAllGood <- sum(totLowGood, totHighGood)
percRxForestHa <-(totRxForest/totRx)*100
percRxForestFRGHa <-(totRxForestFRG/totRx)*100
percRxForestFRGHaOfForest <-(totRxForestFRG/totRxForest)*100
percRxBroadcastOfForestHa <- (totRxForestBroadcast / totRxForest) * 100
percRxBroadcastOfForestFRGHa <- (totRxForestBroadcastFRG / totRxForestFRG) * 100
percRxBroadcasInMatchFRG <- (totRxForestBroadcastFRG / totRxForestBroadcast) * 100


#Percentages


#Summary notes
nfires <- length(unique(mtbs$Event_ID)) + length(unique((smallFires$OBJECTID)))
nForestFires <- length(unique(mtbsForest$Event_ID)) + length(unique((smallFiresForest$OBJECTID)))
nForestFRGLowFires <- length(unique(mtbsForestLowFRG$Event_ID))
rxTypes <- unique(nfporsWithGeeForest$TYPENAME)
nNFPORS <- length(unique(nfporsWithGee$objectid__))
nNFPORSForest <- length(unique(nfporsWithGeeForest$objectid__))
nNFPORSForestLowMix <- length(unique(nfporsWithGeeForestFRGLowMix$objectid__))
percRxForest <- (nNFPORSForest/nNFPORS)*100
percRxForestFRG <- (nNFPORSForestLowMix/nNFPORS)*100


#Tabular stats
summTab1 <- as.data.frame(cbind(c('Low severity good forest fire', 'High severity good forest fire', 'Total any severity good forest fire'), c(totLowGood, totHighGood,totAllGood))) |>
  `names<-`(c('Category', 'TotalHa')) |>
  dplyr::mutate(TotalHa = as.numeric(TotalHa),
                PercForestBurned = TotalHa / totForFire * 100,
                PercForestRx = TotalHa / totRxForest * 100,
                PercForestFRGRx = TotalHa / totRxForestFRG * 100)

summTab2 <- as.data.frame(cbind(c('Wildfire events (n)', 'Prescribed burn events (n)', 'Wildfire burned hectares (ha)', 'Prescribed burn hectares (ha)'),
                                c(nfires, nNFPORS, totFire, totRx),
                                c(nForestFires, nNFPORSForest, totForFire, totRxForest),
                                c("Unable to compute", nNFPORSForestLowMix, totLowGood, totRxForestFRG))) |>
  `names<-`(c('Summary stat', 'Total', 'ForestAmt', 'ForestLowFRGAmt')) |>
  dplyr::mutate(Total = as.numeric(Total),
                ForestAmt = as.numeric(ForestAmt),
                ForestLowFRGAmt = as.numeric(ForestLowFRGAmt),
                ForestPercTot = ForestAmt/Total * 100,
                ForestLowFRGPercForest = ForestLowFRGAmt / ForestAmt * 100)


#Create nice tables
summary_stats_table_1 <- kableExtra::kable(summTab1 %>% dplyr::mutate(dplyr::across(c(PercForestBurned, PercForestRx, PercForestFRGRx), ~sprintf("%s%%", round(.)))),
      caption = "Good fire summary stats table 1",
      format.args = list(big.mark = ","),
      align = rep('l', ncol(summTab1)),
      col.names = c("Category",
                    "Total hectares (ha)",
                    "Percent of total forest area burned (%)",
                    "Percent of forested prescribed burns",
                    "Percent of forested prescribed burns in low/mod severity historical regimes")) |>
  column_spec(1:ncol(summTab1), width = "3cm") |>
  kable_classic(full_width = F, html_font = "Calibri") |>
  save_kable_workaround(here::here(figsDir, 'summary_stats_table_1.png'))

kableExtra::kable(summTab2 %>% dplyr::mutate(dplyr::across(c(ForestPercTot, ForestLowFRGPercForest), ~sprintf("%s%%", round(.)))),
      caption = "Good fire summary stats table 2",
      format.args = list(big.mark = ","),
      align = rep('l', ncol(summTab2)),
      col.names = c("Summary statistic",
                    "Total",
                    "Forested amount (for wildfire events, events w/ >10% forest within perimeter before burn)",
                    "Forested and in low/mod historical regime amount",
                    "Forested amount percent of total",
                    "Forested and in low/mod historical regime percent of forested amount")) |>
  column_spec(1:ncol(summTab2), width = "3cm") |>
  kable_classic(full_width = F, html_font = "Calibri") |>
  save_kable_workaround(here::here(figsDir, 'summary_stats_table_2.png'))


summaryStats <- glue::glue("This analysis included {comma(nfires)} total fire event perimeters, and {comma(nNFPORSForest)} total prescribed forest fire event instances ({comma(nNFPORSForestLowMix)} of those prescribed burn events were in low/mixed severity historical fire regimes). \n\n 'Prescribed fire' was defined as including {paste(rxTypes, collapse = ', ')} management actions. \n\n The total area of low/mod goodfire is {comma(round(totLowGood))} ha, while the total area of high goodfire is {comma(round(totHighGood))} ha. These are respectively {round((totLowGood / totForFire) * 100)}% and {round((totHighGood / totForFire) * 100)}% of the total area of forest burned. High good fire is {round((totHighGood / totLowGood) * 100)}% of the amount of low/mod goodfire. \n\n There were {comma(round(totRxForest))} ha of prescribed forest fires ({comma(round(totRxForestFRG))} ha in low/mixed historical fire regimes). Low/mod goodfire is {round(((totLowGood - totRxForest) / totRxForest) * 100)}% more by area than what was accomplished by prescribed burning ({round(((totLowGood - totRxForestFRG) / totRxForestFRG) * 100)}% more than rx in low/mix). High goodfire is {round((totHighGood / totRxForest) * 100)}% of the area accomplished by prescribed burning ({round((totHighGood / totRxForestFRG) * 100)}% of rx in low/mix). \n\n If we combine high and low/mod good fire, we get {comma(round(totAllGood))} ha of total good fire, which is {round((totAllGood / totForFire) * 100)}% of the total area of forest burned, and {round(((totAllGood - totRxForest) / totRxForest) * 100)}% more by area than what was accomplished by prescribed burning ({round(((totAllGood - totRxForestFRG) / totRxForestFRG) * 100)}% more than rx in low/mix)")
print(summaryStats)
sink(here::here(figsDir, "summary.txt"))
cat(glue::glue("Analysis most recent run: {Sys.time()} \n\n\n"))
cat(summaryStats)
sink()


```



# Combine rx & low-CBI from Rud

```{r}

#STATE SUMMARY

stateAll <- nfporsWithGeeForest |>
  dplyr::group_by(STATEABBR) |>
  dplyr::summarise(rxBurnHa = sum(TOTALACCOMPLISHMENT_HA)) |>
  dplyr::left_join(stateSummary, join_by(STATEABBR == STUSPS))


#YEAR SUMMARY

yearAll <- nfporsWithGeeForest |>
  dplyr::group_by(ACTUALCOMPLETIONYEARNEW) |>
  dplyr::summarise(rxBurnHa = sum(TOTALACCOMPLISHMENT_HA)) |>
  dplyr::left_join(yearSummary, join_by(ACTUALCOMPLETIONYEARNEW == year)) |>
  dplyr::rename(year = ACTUALCOMPLETIONYEARNEW) |>
  dplyr::filter(year < 2021) |>
  dplyr::mutate(year = as.character(year))

#A specific version of pivot_longer to do what is specifically required here
pivot.longer <- function(dat) {
  out <- dat |>
    dplyr::rename(goodFire = `Low/Mod Goodfire (CBI Low/mod in FRG low/mod, HA)`,
                  burnFor = `burnFor (low+mod+high CBI, HA)`) %>%
    {if("% low/mod that is goodfire" %in% names(.)) { 
      dplyr::select(., -`% low/mod that is goodfire`, -`% high that is goodfire`, -`% low/mod/high that is goodfire`)
    } else . } %>%
    {if("Low/Mod CBI (HA)" %in% names(.)) { 
      dplyr::select(., -`Low/Mod CBI (HA)`, -`High CBI (HA)`, -`High Goodfire (CBI High FRG Replacement, HA)`)
    } else . } |>
    tidyr::pivot_longer(cols = c(rxBurnHa,
                                 goodFire,
                                 burnFor),
                        names_to =  "datNm",
                        values_to = "hectares") |>
    dplyr::mutate(datNm = factor(datNm, levels = c("rxBurnHa", "goodFire", "burnFor")))
  return(out)
}

#Use function
yearAllLong <- yearAll |>
  pivot.longer() |>
  dplyr::mutate(acres = hectares * 2.47105)
stateAllLong <- stateAll |>
  pivot.longer() |>
  dplyr::mutate(acres = hectares * 2.47105)
allLong <- yearAll |>
  dplyr::summarise(rxBurnHa = sum(rxBurnHa),
                   `Low/Mod Goodfire (CBI Low/mod in FRG low/mod, HA)` = sum(`Low/Mod Goodfire (CBI Low/mod in FRG low/mod, HA)`),
                   `burnFor (low+mod+high CBI, HA)` = sum(`burnFor (low+mod+high CBI, HA)`)) |>
  pivot.longer() |>
  dplyr::mutate(hectares = round(hectares)) |>
  dplyr::mutate(datNmFull = dplyr::case_when(datNm == 'goodFire' ~ "Good Forest Wildfire",
                                             datNm == 'rxBurnHa' ~ 'Prescribed \nForest Burns',
                                             datNm == 'burnFor' ~ 'Total Forest Wildfire')) |>
  dplyr::mutate(datNm = factor(datNm, levels = c( "burnFor", "goodFire", "rxBurnHa"))) |>
  dplyr::mutate(acres = hectares * 2.47105)

```

# Graphics

```{r}

# Function to get a hex color from R color
# PARAMETERS
# color : a standard R color (e.g. 'red' or 'steelblue')
col2hex <- function(color) {
  rgb_values <- col2rgb(color)
  hex_color <- rgb(rgb_values[1], rgb_values[2], rgb_values[3], maxColorValue=255)
  return(hex_color)
}

#Set graphic parameters
rxCol <- col2hex("darkblue")
gFCol <- "#dc0ab4ff"
totCol <- col2hex("goldenrod2")
plotColors <- c(rxCol, gFCol, totCol)
plotLabels <- c("Prescribed Burns", "Good Wildfire", "Total Wildfire")

#Year graphic
yearGraphicHA <- ggplot2::ggplot(yearAllLong)  +
  geom_col(aes(x = year, y = hectares, fill = datNm),
           position = "dodge", show.legend = TRUE) +
  xlab("Year") +
  ylab("Hectares") +
  labs(title = "Good fire in Western U.S. forests over time (conservative estimate)") +
  scale_y_continuous(labels=comma,
                     limits = c(0, 2000000)#,
                     # sec.axis = sec_axis(trans = ~ . * 2.47105, #add second axis showing acres
                     #                     name = "Acres",
                     #                     labels = comma)
                     ) +
  theme_light() +
  theme(legend.title = element_blank(), legend.position = "bottom",
        text = element_text(size = 15)) + 
  scale_fill_manual(values = plotColors, labels = plotLabels)


#Year graphic
yearGraphicAC <- ggplot2::ggplot(yearAllLong)  +
  geom_col(aes(x = year, y = acres, fill = datNm),
           position = "dodge", show.legend = TRUE) +
  xlab("Year") +
  ylab("Acres") +
  labs(title = "Good fire in Western U.S. forests over time (conservative estimate)") +
  scale_y_continuous(labels=comma,
                     limits = c(0, 5000000)#,
                     # sec.axis = sec_axis(trans = ~ . * 2.47105, #add second axis showing acres
                     #                     name = "Acres",
                     #                     labels = comma)
                     ) +
  theme_light() +
  theme(legend.title = element_blank(), legend.position = "bottom",
        text = element_text(size = 15)) + 
  scale_fill_manual(values = plotColors, labels = plotLabels)

ggplot2::ggsave(plot = yearGraphicHA, filename = here::here(figsDir, 'cbi_rx_year_fig_v3_HA.png'), units = "in", width = 8, bg = "white")
ggplot2::ggsave(plot = yearGraphicAC, filename = here::here(figsDir, 'cbi_rx_year_fig_v3_AC.png'), units = "in", width = 8, bg = "white")

#State graphic
stateGraphicHA <- ggplot2::ggplot(stateAllLong)  +
  geom_col(aes(x = STATEABBR, y = hectares, fill = datNm),
            position = "dodge", show.legend = FALSE) +
  xlab("State") +
  ylab("Hectares") +
  labs(title = "Good fire in Western U.S. forests by state (conservative estimate)") +
  scale_y_continuous(labels=comma,
                     limits = c(0, 2000000)#,
                     # sec.axis = sec_axis(trans = ~ . * 2.47105, #add second axis showing acres
                     #                     name = "Acres",
                     #                     labels = comma)
                     ) +
  theme_light() +
  theme(legend.title = element_blank(),
        text = element_text(size = 15)) + 
  scale_fill_manual(values = plotColors, labels = plotLabels)

stateGraphicAC <- ggplot2::ggplot(stateAllLong)  +
  geom_col(aes(x = STATEABBR, y = acres, fill = datNm),
            position = "dodge", show.legend = FALSE) +
  xlab("State") +
  ylab("Acres") +
  labs(title = "Good fire in Western U.S. forests by state (conservative estimate)") +
  scale_y_continuous(labels=comma,
                     limits = c(0, 5000000)#,
                     # sec.axis = sec_axis(trans = ~ . * 2.47105, #add second axis showing acres
                     #                     name = "Acres",
                     #                     labels = comma)
                     ) +
  theme_light() +
  theme(legend.title = element_blank(),
        text = element_text(size = 15)) + 
  scale_fill_manual(values = plotColors, labels = plotLabels)

ggplot2::ggsave(plot = stateGraphicHA, filename = here::here(figsDir, 'cbi_rx_states_fig_v3_HA.png'), units = "in", width = 8, bg = "white")
ggplot2::ggsave(plot = stateGraphicAC, filename = here::here(figsDir, 'cbi_rx_states_fig_v3_AC.png'), units = "in", width = 8, bg = "white")

#Overall graphic
textSz <- 5.5
#Needs different parameters since turned sideways
plotColors2 <- c(totCol, gFCol, rxCol)
plotLabels2 <- c("Good Forest Wildfire", "Prescribed Forest Burns", "Total Forest Wildfire")

overallGraphicHA <- ggplot2::ggplot(allLong)  +
  geom_col(aes(x = datNm, y = hectares, fill = datNm), position = "dodge", show.legend = FALSE) +
  xlab("Year") +
  ylab("Hectares") +
  #labs(title = "      2010-2020 Good Fire in Western U.S. Forests (conservative estimate)") +
  scale_y_continuous(labels=comma) +
  theme_void() +
  #theme(plot.title = element_text(hjust = 0.5)) + #make title centered
  scale_fill_manual(values = plotColors2, labels = plotLabels2) +
  geom_text(data = allLong |> filter(datNm != "burnFor"),
            aes(x = datNm, y = hectares,
                #label = glue::glue("{scales::comma(hectares)} ha\n({scales::comma(acres)} acres)")),
                label = glue::glue("{scales::comma(hectares)} ha")),
            hjust = -0.1, colour = "black", size = textSz) +
  geom_text(data = allLong |> filter(datNm == "burnFor"),
            aes(x = datNm, y = hectares,
                #label = glue::glue("{scales::comma(hectares)} ha\n({scales::comma(acres)} acres)")),
                label = glue::glue("{scales::comma(hectares)} ha")),
            hjust = 1.2, colour = "white", size = textSz) +
  geom_text(aes(x = datNm, y = 0, label = datNmFull), hjust = -0.1, colour = "white", size = textSz) +
  coord_flip()

overallGraphicAC <- ggplot2::ggplot(allLong)  +
  geom_col(aes(x = datNm, y = acres, fill = datNm), position = "dodge", show.legend = FALSE) +
  xlab("Year") +
  ylab("Acres") +
  #labs(title = "      2010-2020 Good Fire in Western U.S. Forests (conservative estimate)") +
  scale_y_continuous(labels=comma) +
  theme_void() +
  #theme(plot.title = element_text(hjust = 0.5)) + #make title centered
  scale_fill_manual(values = plotColors2, labels = plotLabels2) +
  geom_text(data = allLong |> filter(datNm != "burnFor"),
            aes(x = datNm, y = acres,
                #label = glue::glue("{scales::comma(hectares)} ha\n({scales::comma(acres)} acres)")),
                label = glue::glue("{scales::comma(acres)} acres")),
            hjust = -0.1, colour = "black", size = textSz) +
  geom_text(data = allLong |> filter(datNm == "burnFor"),
            aes(x = datNm, y = acres,
                #label = glue::glue("{scales::comma(hectares)} ha\n({scales::comma(acres)} acres)")),
                label = glue::glue("{scales::comma(acres)} acres")),
            hjust = 1.2, colour = "white", size = textSz) +
  geom_text(aes(x = datNm, y = 0, label = datNmFull), hjust = -0.1, colour = "white", size = textSz) +
  coord_flip()



ggplot2::ggsave(plot = overallGraphicAC, filename = here::here(figsDir, 'overall_v2_AC.png'), units = "in", width = 6, height = 2.5, bg = "transparent")
ggplot2::ggsave(plot = overallGraphicHA, filename = here::here(figsDir, 'overall_v2_HA.png'), units = "in", width = 6, height = 2.5, bg = "transparent")


####COMBINED GRAPHIC

# #Arranging complex grobs: https://cran.r-project.org/web/packages/gridExtra/vignettes/arrangeGrob.html
# lay <- rbind(c(1, NA),
#              c(2,3))
# 
# combinedPlot <- gridExtra::arrangeGrob(overallGraphic,
#                                        stateGraphic + ggplot2::labs(title = NULL),
#                                        yearGraphic + ggplot2::labs(title = NULL),
#                                        layout_matrix = lay)#,
#                                        #top = "Good Fire & Rx in the Western US (conservative estimate)")
# ggplot2::ggsave(combinedPlot, filename = here::here(figsDir, 'combined_fig1.png'), units = "in", width = 10)
# 


library(patchwork)

combinedPlot2HA <- (stateGraphicHA + ggplot2::labs(title = NULL)) / (yearGraphicHA + ggplot2::labs(title = NULL))
ggplot2::ggsave(combinedPlot2HA, filename = here::here(figsDir, 'combined_fig2_HA.png'), units = "in", width = 6, height = 8)

combinedPlot2AC <- (stateGraphicAC + ggplot2::labs(title = NULL)) / (yearGraphicAC + ggplot2::labs(title = NULL))
ggplot2::ggsave(combinedPlot2AC, filename = here::here(figsDir, 'combined_fig2_AC.png'), units = "in", width = 6, height = 8)





```

# Tables

```{r}

#Prep forest numbers
stateForest <- fPixAll |>
  dplyr::group_by(NAME) |>
  dplyr::summarise(forestHectares = mean(hectares)) |>
  dplyr::rename(Group = NAME) |>
  dplyr::ungroup()
yearForest <- fPixAll |>
  dplyr::group_by(year) |>
  dplyr::summarise(forestHectares = sum(hectares)) |>
  dplyr::rename(Group = year) |>
  dplyr::ungroup()
allForest <- rbind(stateForest, yearForest) |>
  rbind(c("All Years", mean(yearForest$forestHectares))) |>
  dplyr::mutate(forestHectares = round(as.numeric(forestHectares)))

#Get table with overall numbers
overallTblHA <- stateAllLong |> 
  dplyr::select(-acres) |>
  tidyr::pivot_wider(names_from = "datNm",
                     values_from = "hectares") |>
  dplyr::mutate(percGoodOfTot = paste(round(100 * (goodFire / burnFor)), "%", sep = "")) |>
  #dplyr::mutate(percRxOfTot = paste(round(100 * (rxBurnHa / burnFor)), "%", sep = "")) |>
  dplyr::select(-STATEABBR) |>
  dplyr::rename(Group = NAME) |>
  rbind(
    yearAllLong |>
      dplyr::select(-acres) |>
      tidyr::pivot_wider(names_from = "datNm",
                         values_from = "hectares") |>
      dplyr::mutate(percGoodOfTot = paste(round(100 * (goodFire / burnFor)), "%", sep = "")) |>
      dplyr::rename(Group = year)
  ) |>
  rbind(
    allLong |>
      dplyr::select(-datNmFull, -acres) |>
      tidyr::pivot_wider(names_from = "datNm",
                         values_from = "hectares") |>
      dplyr::mutate(percGoodOfTot = paste(round(100 * (goodFire / burnFor)), "%", sep = ""),
                    Group = "All Years")
  ) |>
  dplyr::mutate(dplyr::across(.cols = rxBurnHa:burnFor, round)) |>
  dplyr::left_join(allForest, by = "Group") |>
  dplyr::mutate(percGoodOfFor = paste(round(100 * (goodFire / forestHectares), 1), "%", sep = "")) |>
  dplyr::relocate(forestHectares, .after = burnFor)



overallTblAC <- stateAllLong |> 
  dplyr::select(-acres) |>
  tidyr::pivot_wider(names_from = "datNm",
                     values_from = "hectares") |>
  dplyr::mutate(percGoodOfTot = paste(round(100 * (goodFire / burnFor)), "%", sep = "")) |>
  #dplyr::mutate(percRxOfTot = paste(round(100 * (rxBurnHa / burnFor)), "%", sep = "")) |>
  dplyr::select(-STATEABBR) |>
  dplyr::rename(Group = NAME) |>
  rbind(
    yearAllLong |>
      dplyr::select(-acres) |>
      tidyr::pivot_wider(names_from = "datNm",
                         values_from = "hectares") |>
      dplyr::mutate(percGoodOfTot = paste(round(100 * (goodFire / burnFor)), "%", sep = "")) |>
      dplyr::rename(Group = year)
  ) |>
  rbind(
    allLong |>
      dplyr::select(-datNmFull, -acres) |>
      tidyr::pivot_wider(names_from = "datNm",
                         values_from = "hectares") |>
      dplyr::mutate(percGoodOfTot = paste(round(100 * (goodFire / burnFor)), "%", sep = ""),
                    Group = "All Years")
  ) |>
  dplyr::mutate(dplyr::across(.cols = rxBurnHa:burnFor,  ~ . * 2.47105)) |>
  dplyr::mutate(dplyr::across(.cols = rxBurnHa:burnFor, round))



#Create nice table
kableExtra::kable(overallTblHA,
      caption = "Good fire in the western U.S. by year, state, and overall (2010-2020)",
      format.args = list(big.mark = ","),
      align = rep('l', 7),
      col.names = c("Group","Prescribed burn forest HA", "Good forest wildfire HA", "Total forest wildfire HA", "Total forest HA", "Good forest wildfire as a percentage of total forest wildfire", "Good forest wildfire as a percentage of total forest")) |>
  kableExtra::column_spec(1, width = "1cm") |>
  kableExtra::column_spec(2:7, width = "3cm") |>
  kableExtra::kable_classic(full_width = F, html_font = "Calibri") |>
  row_spec(11, extra_css = "border-bottom: 1px dashed;") |>
  save_kable_workaround(here::here(figsDir, 'summary_table_v1_HA.png'))

kableExtra::kable(overallTblAC,
      caption = "Good fire in the western U.S. by year, state, and overall (2010-2020)",
      format.args = list(big.mark = ","),
      align = rep('l', 5),
      col.names = c("Group","Prescribed burn forest acres", "Good forest wildfire acres", "Total forest wildfire acres", "Good forest wildfire as a percentage of total forest wildfire")) |>
  kableExtra::column_spec(1, width = "1cm") |>
  kableExtra::column_spec(2:5, width = "3cm") |>
  kableExtra::kable_classic(full_width = F, html_font = "Calibri") |>
  save_kable_workaround(here::here(figsDir, 'summary_table_v1_AC.png'))

```

# Create CA graphics

```{r california-graphics}


#Pull data by-year for CA, calculate totals
#NEW data summaries from Rud, includes small fires
cbiLowAll <- googlesheets4::read_sheet("https://docs.google.com/spreadsheets/d/1wUWLin5cXSd9uflna9f_hliz5HEWxOcc7a58mtDziEk/edit#gid=1307905708", sheet = "CBI lowmod All", range = "A1:N13")
cbiHighAll <- googlesheets4::read_sheet("https://docs.google.com/spreadsheets/d/1wUWLin5cXSd9uflna9f_hliz5HEWxOcc7a58mtDziEk/edit#gid=1307905708", sheet = "CBI high all", range = "A1:N13")
cbiGood <- googlesheets4::read_sheet("https://docs.google.com/spreadsheets/d/1wUWLin5cXSd9uflna9f_hliz5HEWxOcc7a58mtDziEk/edit#gid=1307905708", sheet = "CBI lowmod FRG lowmod", range = "A1:N13")


#A function to get CA stats in long form from Rud's summaries
get.Ca.Stats <- function(dats, fieldnm) {
  out <- dats |>
  dplyr::select(-`Total (HA)`) |>
  dplyr::filter(NAME == "California") |>
  tidyr::pivot_longer(cols = b2010:b2020,
                      names_to = c("year"),
                      values_to = fieldnm)
  return(out)
}


#Create CA by time dataset
#do rx data
caRxTime <- nfporsWithGeeForest |>
  dplyr::filter(STATEABBR == "CA") |>
  dplyr::group_by(STATEABBR, ACTUALCOMPLETIONYEARNEW) |>
  dplyr::summarise(rxBurnHa = sum(TOTALACCOMPLISHMENT_HA)) |>
  dplyr::rename(year = ACTUALCOMPLETIONYEARNEW) |>
  dplyr::ungroup() |>
  dplyr::select(-STATEABBR)

#Merge to dats from Rud
caTimeStatsLong <- cbiLowAll |>
  get.Ca.Stats("CBILow") |>
  dplyr::left_join(cbiHighAll |> get.Ca.Stats("CBIHigh")) |>
  dplyr::left_join(cbiGood |> get.Ca.Stats("goodFire")) |> #merge all the data for CA
  dplyr::mutate(burnFor = CBILow+CBIHigh, #Get burnFor
                year = as.integer(substr(year, 2, nchar(year)))) |>
  dplyr::mutate(burnFor = burnFor*0.0001,
                goodFire = goodFire*0.0001) |> #convert to HA from m^2
  dplyr::select(-CBILow, -CBIHigh, -NAME, -STUSPS) |>
  dplyr::left_join(caRxTime) |>
  tidyr::pivot_longer(cols = c(rxBurnHa,
                             goodFire,
                             burnFor),
                      names_to =  "datNm",
                      values_to = "hectares") |>
  dplyr::mutate(datNm = factor(datNm, levels = c("rxBurnHa", "goodFire", "burnFor")),
                year = as.character(year))
  


#Get overall CA stats for overall summary graphic
caLong <- stateAllLong |>
  dplyr::filter(STATEABBR == "CA") |>
  dplyr::mutate(datNmFull = dplyr::case_when(datNm == 'goodFire' ~ "Good Forest Wildfire",
                                             datNm == 'rxBurnHa' ~ 'Prescribed \nForest Burns',
                                             datNm == 'burnFor' ~ 'Total Forest Wildfire')) |>
  dplyr::mutate(datNm = factor(datNm, levels = c( "burnFor", "goodFire", "rxBurnHa"))) |>
  dplyr::mutate(acres = round(hectares * 2.47105))


#CA summary graphic
  #Needs different parameters since turned sideways
  plotColors2 <- c("gray", "#dc0ab4ff", "seagreen")
  plotLabels2 <- c("Good Forest Wildfire", "Prescribed Forest Burns", "Total Forest Wildfire")
  
caSummary <- ggplot2::ggplot(caLong) +
  geom_col(aes(x = datNm, y = hectares, fill = datNm), position = "dodge", show.legend = FALSE) +
  xlab("Year") +
  ylab("Hectares") +
  labs(title = "      2010-2020 Good Fire in California Forests (conservative estimate)") +
  scale_y_continuous(labels=comma) +
  theme_void() +
  #theme(plot.title = element_text(hjust = 0.5)) + #make title centered
  scale_fill_manual(values = plotColors2, labels = plotLabels2) +
  geom_text(data = caLong |> filter(datNm != "burnFor"), aes(x = datNm, y = hectares, label = glue::glue("{scales::comma(hectares)} ha\n({scales::comma(acres)} acres)")), hjust = -0.1, colour = "black", size = 4) +
  geom_text(data = caLong |> filter(datNm == "burnFor"), aes(x = datNm, y = hectares, label = glue::glue("{scales::comma(hectares)} ha\n({scales::comma(acres)} acres)")), hjust = 1.1, colour = "white", size = 4) +
  geom_text(aes(x = datNm, y = 0, label = datNmFull), hjust = -0.1, colour = "white", size = 4) +
  coord_flip()  
caSummary

ggplot2::ggsave(plot = caSummary, filename = here::here(figsDir, 'california_summary.png'), units = "in", width = 6, height = 2.5, bg = "white")

#Year graphic
plotColors <- c("seagreen", "#dc0ab4ff", "gray")
plotLabels <- c("Prescribed Burns", "Good Wildfire", "Total Wildfire")
caYearGraphic <- ggplot2::ggplot(caTimeStatsLong)  +
  geom_col(aes(x = year, y = hectares, fill = datNm),
           position = "dodge", show.legend = FALSE) +
  xlab("Year") +
  ylab("Hectares") +
  #labs(title = "Good fire in California forests over time (conservative estimate)") +
  scale_y_continuous(labels=comma,
                     sec.axis = sec_axis(trans = ~ . * 2.47105, #add second axis showing acres
                                         name = "Acres",
                                         labels = comma)) +
  theme_light() +
  theme(legend.title = element_blank()) + 
  scale_fill_manual(values = plotColors, labels = plotLabels)
ggplot2::ggsave(plot = caYearGraphic, filename = here::here(figsDir, 'ca_time.png'), units = "in", width = 6, height = 3, bg = "white")




```
